Phill: Fixed various bugs, removed lots of cout bloat that was making debugging difficult. We figured out that the Layer ID is set correctly (now). However there is a problem withe the maximum number of layers not being large enough. Either the maximum number of layers we have is incorrect or the cell ID positions and lengths for finding a layer are incorrect.

+++ include/LayerCodeInCellID.h
@@ -6,7 +6,13 @@
 class LayerCodeInCellID {
 
   public:
-    LayerCodeInCellID(std::string const CellID, int const LayerInfoStart, int const LayerInfoLength) : CellID_(CellID), StartBin_layers(LayerInfoStart), LengthBin_layers(LayerInfoLength), Layer(-1) {}
+    LayerCodeInCellID(std::string const CellID, 
+        int const LayerInfoStart, 
+        int const LayerInfoLength) : 
+      CellID_(CellID), 
+      StartBin_layers(LayerInfoStart), 
+      LengthBin_layers(LayerInfoLength), 
+      Layer(-1) {}
 
     virtual ~LayerCodeInCellID() {}
 
@@ -20,7 +26,7 @@ class LayerCodeInCellID {
     LayerCodeInCellID(float ID): CellID_(""), StartBin_layers(-1), LengthBin_layers(-1), Layer(-1) {}
     LayerCodeInCellID(double ID): CellID_(""), StartBin_layers(-1), LengthBin_layers(-1), Layer(-1) {}
 
-    int Layer;
+    unsigned long Layer;
     int StartBin_layers;
     int LengthBin_layers;
